requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
anthropic==0.34.2
python-docx==1.1.2
pdfplumber==0.11.4
supabase==2.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.2
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
